% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{.mlc_character}
\alias{.mlc_character}
\title{An R scalar character type}
\format{
An object of class \code{character} of length 1.
}
\usage{
.mlc_character
}
\description{
An R scalar character type
}
\keyword{datasets}
