% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{.mlc_integer}
\alias{.mlc_integer}
\title{An R scalar integer type}
\format{
An object of class \code{list} of length 1.
}
\usage{
.mlc_integer
}
\description{
An R scalar integer type
}
\keyword{datasets}
