% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{.mlc_list}
\alias{.mlc_list}
\title{Helper function for building list types}
\usage{
.mlc_list(x)
}
\arguments{
\item{x}{The type parameter for a list}
}
\description{
The "list" here is a homogenous vector, like the Haskell or lisp list, not
the named, heterogenous lists of R. So `mlc_list(mlc_integer)` would be the
R version of the Haskell type `[Int]`, in this case, simply a numeric
vector.
}
\examples{
# parse a JSON list into R
mlc_deserialize("[1,2,3]", .mlc_list(.mlc_integer))

# write an R vector to JSON
mlc_serialize(c(1,2,3), .mlc_list(.mlc_integer))
}
