% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{.mlc_numeric}
\alias{.mlc_numeric}
\title{An R scalar numeric type}
\format{
An object of class \code{character} of length 1.
}
\usage{
.mlc_numeric
}
\description{
An R scalar numeric type
}
\keyword{datasets}
