% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{.mlc_tuple}
\alias{.mlc_tuple}
\title{Helper function for building tuple types}
\usage{
.mlc_tuple(...)
}
\arguments{
\item{x}{The type parameter for a list}
}
\description{
A tuple corresponds to an unnamed R list.
}
\examples{
# parse a JSON list into R
deserialize("[1,2,3]", mlc_tuple(mlc_integer))

# write an R list to JSON
serialize(list(1,2,3), mlc_tuple(mlc_integer))
}
