\name{rb.excludeTables}
\alias{rb.excludeTables}
\title{Exclude some table names}
\usage{
  rb.excludeTables(tablename,
    exclude = options("rb.excludes"))
}
\arguments{
  \item{tablename}{is the proposed tablename}

  \item{exclude}{a regular expression of exclusions}
}
\description{
  Exclude some table names
}

