% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{serialize}
\alias{serialize}
\title{Serialize R data using a JSON type string}
\usage{
serialize(x, schema)
}
\arguments{
\item{x}{R data of (nearly) any type}

\item{schema}{JSON type string}
}
\description{
Serialize R data using a JSON type string
}
